requirements.txt
apns2==0.7.1
asgiref==3.2.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.4
django-cors-headers==3.2.1
django-push-notifications==2.0.0
djangorestframework==3.11.0
fcm-django==0.3.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
py-vapid==1.7.0
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
pytz==2019.3
pywebpush==1.10.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
